+++ jaffle/command/start/buffer.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+
+class LogBuffer(object):
+
+    def __init__(self):
+        pass