+++ jaffle/config/jaffle_config.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
 import hcl
+import json
+import jsonschema
 from pathlib import Path
 import re
 from ..functions import functions
@@ -53,6 +55,16 @@ class JaffleConfig(object):
             app_name: [(re.compile(str_value(r['from'])), r['to']) for r in
                        app_data.get('logger', ConfigDict()).get('replace_regex', ConfigList())]
             for app_name, app_data in self.app.items()
+        }
+        self.process_log_suppress_patterns = {
+            app_name: [re.compile(str_value(r)) for r in
+                       app_data.get('logger', ConfigDict()).get('suppress_regex', ConfigList())]
+            for app_name, app_data in self.process.items()
+        }
+        self.process_log_replace_patterns = {
+            app_name: [(re.compile(str_value(r['from'])), r['to']) for r in
+                       app_data.get('logger', ConfigDict()).get('replace_regex', ConfigList())]
+            for app_name, app_data in self.process.items()
         }
         self.global_log_suppress_patterns = [
             re.compile(str_value(r))
@@ -107,8 +119,13 @@ class JaffleConfig(object):
         runtime_variables : dict
             Runtime variables.
         """
-        return cls.create(deep_merge(*(cls._load_file(f) for f in file_paths)),
-                          raw_namespace, runtime_variables)
+        data = deep_merge(*(cls._load_file(f) for f in file_paths))
+
+        with (Path(__file__).parent.parent / 'schema' / 'config_schema.json').open() as sf:
+            schema = json.load(sf)
+        jsonschema.validate(data, schema)
+
+        return cls.create(data, raw_namespace, runtime_variables)
 
     @classmethod
     def create(cls, data_dict, raw_namespace, runtime_variables):